+++ packages/codemod/src/transformations/v4/authProvider-to-legacyAuthProvider.ts
