+++ packages/codemod/src/transformations/v4/separate-imports-react-hook-form.ts
@@ -1,10 +1,10 @@
 import { API, FileInfo } from "jscodeshift";
 import fs from "fs";
 import path from "path";
-import { addPackage, install, isPackageJsonUpdated } from "../helpers";
-import checkPackageLock from "../helpers/checkPackageLock";
-import separateImports from "../helpers/separateImports";
-import { exported } from "../definitions/separated-imports/react-hook-form";
+import { addPackage, install, isPackageJsonUpdated } from "../../helpers";
+import checkPackageLock from "../../helpers/checkPackageLock";
+import separateImports from "../../helpers/separateImports";
+import { exported } from "../../definitions/separated-imports/react-hook-form";
 
 export const parser = "tsx";
